refactor(internal): share cache-line padding between primitives
