+++ asyncband/src/internal/mod.rs
@@ -70,6 +70,9 @@ pub(crate) mod atomic_waker;
 #[allow(dead_code)]
 pub(crate) mod arena;
 
+#[allow(dead_code)]
+pub(crate) mod cache_padded;
+
 #[cfg(any(feature = "latch", feature = "once"))]
 pub(crate) mod countdown;
